implements feedback
