+++ examples/with-emailverification-with-otp/README.md
@@ -1,6 +1,6 @@
 ![SuperTokens banner](https://raw.githubusercontent.com/supertokens/supertokens-logo/master/images/Artboard%20%E2%80%93%2027%402x.png)
 
-# SuperTokens EmailPassword Demo app with Email Verification through OTP
+# SuperTokens EmailPassword and Social Login Demo app with Email Verification through OTP
 
 This demo app demonstrates the following use cases:
 
@@ -45,9 +45,9 @@ You can take a look at our [quick setup](https://supertokens.com/docs/thirdparty
 
 Since Email Verification is enabled when a user signs up they will be redirected to a new screen which will prompt them to go to their email and click on the email verification link.
 
-We will need to override the component which handles this flow to use our own custom UI which will allow the user to enter an OTP. We can do so by overriding the `EmailVerificationSendVerifyEmail_Override` component in our frontend.
+We will need to override the component which handles this flow to use our own custom component with UI to allow the user to enter an OTP. We can do so by overriding the `EmailVerificationSendVerifyEmail_Override` component in our frontend.
 
-Our custom component will contain the UI to allow the user to enter an OTP. It will also have functions that will query our backend server's APIs to send the OTP email to the user and verify its email with the OTP.
+Our custom component will also have functions that will query our backend server's APIs to send the OTP email to the user and verify its email with the OTP.
 
 ![EmailVerification with OTP UI](./images/emailverification-with-otp.png)
 
@@ -60,13 +60,19 @@ Our backend will need to handle two main things:
 
 #### Generating the OTP and sending the email to the user
 
+improvements
 In the regular email verification flow, the url which is sent to the user for verification contains a token. This token will be used for verifying the user.
 
 We will use the `createAndSendCustomEmail` function in our backend config where we will generate an OTP, map it to the token and send the OTP to the user via email using node mailer.
 
 #### Verifying the OTP
 
-The `verifyEmailPOST` api handles email verification on the backend. We will need to override the default behavior and check that the OTP sent is mapped to a token. If a mapping exists we can retrieve the token and verify the users email; otherwise we pass the otp as the token which will result in the email verification failing.
+The `verifyEmailPOST` api handles email verification on the backend. We will need to override the default behavior and check that the OTP sent is mapped to a token. If a mapping exists we can retrieve the token and verify the users email; if no mapping exists we return an invalid token error.
+
+## Future improvements
+
+-   In our example app we are storing the OTP to token mapping in memory. Ideally you would store this mapping in a database for persistant storage.
+-   Currently OTPs will continue to remain in the mapping even after use. An additional improvement could be to create a cronjob which would periodically remove old OTPs.
 
 This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
